multi: drop sqlite
Upgrade the pgsql database, adding a stats table and a winners column
to the blocks table. This is a lengthy (15-30 min) upgrade that requires
pulling MsgBlocks for the entire blockchain and evolving the ticket pool.

* migrate most of extended_stake table

* migrate API and Explorer

* transfer stakeDB management.

* sequence table creation (not random order)

* combine ChainDB and ChainDBRPC

* add missing atoms conversion in chart controller

+++ api/types/apitypes.go
@@ -579,7 +579,7 @@ type TicketPoolValsAndSizes struct {
 	StartHeight uint32    `json:"start_height"`
 	EndHeight   uint32    `json:"end_height"`
 	Value       []float64 `json:"value"`
-	Size        []float64 `json:"size"`
+	Size        []uint32  `json:"size"`
 }
 
 // BlockDataBasic models primary information about a block.
